Some changes to the UI and regexp.
The Ask function will now only take a newline before
getting input if the multiline flag has been set - I
feel that this is an UI enhancement.

If the user specifies an invalid flag they are now told
which flag was invalid, and shown the help output.

The username regexp now only matches usernames of the
form "s[numbers]", as opposed to before when it would
match any string with "s[numbers]" in it. Similarily,
the urgent flag will only accept the inputs 1 or 0,
not any string with a 1 or 0 in it.