Proper ARGV checking, and thread-check removal.
Findme now checks for the number of ARGVs correctly.
The mid-way clearing of the threads was also removed,
which speeds up the system by about 50% - it takes
about 25 seconds to search all of AT. However, there
have been problems in the past with that many threads
causing fork error messages. We will have to monitor
that situation.